% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm for GCD.}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{numeric.}

\item{b}{numeric.}
}
\value{
If input data is numeric or integer then 
the output will be a integer otheriwise it will return an error.
}
\description{
The function returns the greatest common divisor of two numbers.
}
\references{
\href{https://en.wikipedia.org/wiki/Euclideanalgorithm}{Euclideanalgorithm}
}
