\name{Euclistra-package}
\alias{Euclistra-package}
\alias{Euclistra}
\docType{package}
\title{
\packageTitle{Euclistra}
}
\description{
\packageDescription{Euclistra}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Euclistra}
\packageIndices{Euclistra}
}
\author{
\packageAuthor{Euclistra}
Maintainer: \packageMaintainer{Euclistra}
}
\keyword{ package }


