\name{Euclistra-package}
\alias{Euclistra-package}
\alias{Euclistra}
\docType{package}
\title{
\packageTitle{Euclistra}
}
\description{
\packageDescription{Euclistra}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Euclistra}
\packageIndices{Euclistra}
An overview of how to use the package, including the
most important functions
}
\author{
\packageAuthor{Euclistra}

Maintainer: \packageMaintainer{Euclistra}
}
\references{
Literature or other references for background information
}
Optionally other standard keywords, one per line, from
file KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
\code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
